Fixing issues around queries

+++ scripts/setup.sql
@@ -3,7 +3,7 @@ CREATE DATABASE IF NOT EXISTS xspends;
 USE xspends;
 
 CREATE TABLE IF NOT EXISTS `users` (
-    `id` INT NOT NULL,
+    `id` BIGINT NOT NULL,
     `username` VARCHAR(255) NOT NULL UNIQUE,
     `name` VARCHAR(255) NOT NULL,
     `email` VARCHAR(255) NOT NULL UNIQUE,
@@ -15,8 +15,8 @@ CREATE TABLE IF NOT EXISTS `users` (
 );
 
 CREATE TABLE IF NOT EXISTS `categories` (
-    `id` INT NOT NULL,
-    `user_id` INT NOT NULL,
+    `id` BIGINT NOT NULL,
+    `user_id` BIGINT NOT NULL,
     `name` VARCHAR(255) NOT NULL,
     `description` TEXT,
     `icon` VARCHAR(255),
@@ -39,8 +39,8 @@ CREATE TABLE IF NOT EXISTS `sources` (
 );
 
 CREATE TABLE IF NOT EXISTS `tags` (
-    `id` INT NOT NULL,
-    `user_id` INT NOT NULL,
+    `id` BIGINT NOT NULL,
+    `user_id` BIGINT NOT NULL,
     `name` VARCHAR(255) NOT NULL UNIQUE,
     `created_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `updated_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
@@ -49,13 +49,13 @@ CREATE TABLE IF NOT EXISTS `tags` (
 );
 
 CREATE TABLE IF NOT EXISTS `transactions` (
-    `id` INT NOT NULL,
-    `user_id` INT NOT NULL,
+    `id` BIGINT NOT NULL,
+    `user_id` BIGINT NOT NULL,
     `type` VARCHAR(255) NOT NULL DEFAULT 'SAVINGS',
-    `source_id` INT,
+    `source_id` BIGINT,
     `amount` DECIMAL(10, 2) NOT NULL,
     `timestamp` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    `category_id` INT,
+    `category_id` BIGINT,
     FOREIGN KEY (`user_id`) REFERENCES `users`(`id`),
     FOREIGN KEY (`source_id`) REFERENCES `sources`(`id`),
     FOREIGN KEY (`category_id`) REFERENCES `categories`(`id`),
@@ -64,8 +64,8 @@ CREATE TABLE IF NOT EXISTS `transactions` (
 
 -- Create a transaction_tags junction table to store the many-to-many relationship between transactions and tags
 CREATE TABLE IF NOT EXISTS `transaction_tags` (
-    `transaction_id` INT NOT NULL,
-    `tag_id` INT NOT NULL,
+    `transaction_id` BIGINT NOT NULL,
+    `tag_id` BIGINT NOT NULL,
     `created_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `updated_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     FOREIGN KEY (`transaction_id`) REFERENCES `transactions`(`id`),
